删除微博

+++ app/Http/Controllers/StatusesController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use Auth;
+use App\Models\Status;
 
 class StatusesController extends Controller
 {
@@ -31,8 +32,10 @@ class StatusesController extends Controller
     // 删除微博
     public function destroy(Status $status)
     {
-        $status->delete();
 
+        $this->authorize('destroy', $status);
+
+        $status->delete();
         session()->flash('success', '删除成功！');
         return redirect()->back();
     }
